+++ packages/pash-prompt/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const { PromptEngine, generateSystemPrompt } = require('./src/engine');
+
+module.exports = { PromptEngine, generateSystemPrompt };